run.py

+++ run.py
@@ -1,6 +1,6 @@
+from app.helpers import utils
 from app import create_app, db
 from app.models import WikiData
-from app.helpers import utils
 from datetime import datetime as dt
 
 app = create_app()
@@ -11,7 +11,7 @@ if __name__ == "__main__":
         records = WikiData.query.all()
         current_date = dt.now().strftime("%d-%m")
         if len(records) == 0 or records[0].date[:5] != current_date:
-            if records[0].date[:5] != current_date:
+            if len(records) != 0 and records[0].date[:5] != current_date:
                 for record in records:
                     db.session.delete(record)
                 db.session.commit()
